warning removal?

+++ fork/os/process.c
@@ -1188,11 +1188,11 @@ The items that need to be fixed for this lab in DLXOS are:
   bcopy((char *)currentPCB->currentSavedFrame[PROCESS_STACK_PTBASE], (char *)newSystemStackPage, MEM_PAGESIZE);
 
   childPCB->currentSavedFrame = (childPCB->currentSavedFrame && 0x00000FFF) | newSystemStackPage;
-  dbprintf('m', "ProcessRealFork(%d): updated childPCB currentSavedFrame to 0x%x.\n", GetCurrentPid(), childPCB->currentSavedFrame);
+  dbprintf('m', "ProcessRealFork(%d): updated childPCB currentSavedFrame to 0x%x.\n", GetCurrentPid(), (uint32)childPCB->currentSavedFrame);
   childPCB->currentSavedFrame[PROCESS_STACK_PTBASE] = (childPCB->currentSavedFrame[PROCESS_STACK_PTBASE] && 0x00000FFF) | newSystemStackPage;
-  dbprintf('m', "ProcessRealFork(%d): updated childPCB PTBase to 0x%x.\n", GetCurrentPid(), childPCB->currentSavedFrame[PROCESS_STACK_PTBASE]);
+  dbprintf('m', "ProcessRealFork(%d): updated childPCB PTBase to 0x%x.\n", GetCurrentPid(), (uint32)childPCB->currentSavedFrame[PROCESS_STACK_PTBASE]);
   childPCB->sysStackPtr = (childPCB->sysStackPtr && 0x00000FFF) | newSystemStackPage;
-  dbprintf('m', "ProcessRealFork(%d): updated childPCB sysStackPtr to 0x%x.\n", GetCurrentPid(), childPCB->sysStackPtr);
+  dbprintf('m', "ProcessRealFork(%d): updated childPCB sysStackPtr to 0x%x.\n", GetCurrentPid(), (uint32)childPCB->sysStackPtr);
 
   for (i = 0; i < MEM_NUM_PAGE_TABLE_ENTRIES; i++)
   {
